Form data

+++ src/main/java/contracts/repository/AccountRepository.java
@@ -1,7 +1,8 @@
 package contracts.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import contracts.domain.user;
 
-public interface AccountRepository extends JpaRepository<user, Long> {
+public interface AccountRepository extends JpaRepository<user, Integer> {
 
 }
